move ParseRPCSerialPortSettings and InitPort methods to rpc_helpers
