+++ src/rpc/rpc_helpers.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "serial_port_settings.h"
+#include "serial_driver.h"
 #include <wblib/json_utils.h>
 
 TSerialPortConnectionSettings ParseRPCSerialPortSettings(const Json::Value& request);
+PPort InitPort(const Json::Value& request);
