+++ src/rpc/rpc_port_driver_list.cpp
@@ -1,7 +1,5 @@
 #include "rpc_port_driver_list.h"
-#include "rpc_serial_port_settings.h"
-#include "serial_port.h"
-#include "tcp_port.h"
+#include "rpc_helpers.h"
 
 #define LOG(logger) ::logger.Log() << "[RPC] "
 
@@ -45,27 +43,4 @@ PRPCPortDriver TRPCPortDriverList::Find(const Json::Value& request) const
         default:
             throw TRPCException("More than one matches for requested port", TRPCResultCode::RPC_WRONG_PORT);
     }
-}
-
-PPort TRPCPortDriverList::InitPort(const Json::Value& request)
-{
-    if (request.isMember("path")) {
-        std::string path;
-        WBMQTT::JSON::Get(request, "path", path);
-        TSerialPortSettings settings(path, ParseRPCSerialPortSettings(request));
-
-        LOG(Debug) << "Create serial port: " << path;
-        return std::make_shared<TSerialPort>(settings);
-    }
-    if (request.isMember("ip") && request.isMember("port")) {
-        std::string address;
-        int portNumber = 0;
-        WBMQTT::JSON::Get(request, "ip", address);
-        WBMQTT::JSON::Get(request, "port", portNumber);
-        TTcpPortSettings settings(address, portNumber);
-
-        LOG(Debug) << "Create tcp port: " << address << ":" << portNumber;
-        return std::make_shared<TTcpPort>(settings);
-    }
-    throw TRPCException("Port is not defined", TRPCResultCode::RPC_WRONG_PARAM_VALUE);
 }
