+++ src/rpc/rpc_port_driver_list.h
@@ -17,7 +17,6 @@ class TRPCPortDriverList
 public:
     TRPCPortDriverList(PRPCConfig rpcConfig, PMQTTSerialDriver serialDriver);
     PRPCPortDriver Find(const Json::Value& request) const;
-    PPort InitPort(const Json::Value& request);
 
 private:
     std::vector<PRPCPortDriver> Drivers;
